Move database file into utils
